# backend/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Backend API for RockyDex Mobile App, supporting user authentication, favorites, and reading history synchronization.",
    "title": "RockyDex API",
    "contact": {},
    "version": "1.1.2"
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/user/profile": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Returns the user's profile information including preferences.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "user"
        ],
        "summary": "Get user profile",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/domain.User"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Updates user profile avatar, theme, layout, or brightness settings.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "user"
        ],
        "summary": "Update user profile preferences",
        "parameters": [
          {
            "description": "Profile details to update",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.UpdateProfileInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/domain.Profile"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/user/stats": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Returns counts of read comics, favorites, and chapters read.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "user"
        ],
        "summary": "Get user reading statistics",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.ReadingStats"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "dto.UpdateProfileInput": {
      "type": "object",
      "properties": {
        "avatar_url": {
          "type": "string"
        },
        "theme_preference": {
          "type": "string"
        },
        "reading_layout": {
          "type": "string"
        },
        "reading_brightness": {
          "type": "number"
        }
      }
    },
    "dto.ReadingStats": {
      "type": "object",
      "properties": {
        "total_comics_read": {
          "type": "integer"
        },
        "total_favorites": {
          "type": "integer"
        },
        "chapters_read": {
          "type": "integer"
        }
      }
    },
    "domain.User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        },
        "profile": {
          "$ref": "#/definitions/domain.Profile"
        }
      }
    },
    "domain.Profile": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "string"
        },
        "avatar_url": {
          "type": "string"
        },
        "theme_preference": {
          "type": "string"
        },
        "reading_layout": {
          "type": "string"
        },
        "reading_brightness": {
          "type": "number"
        },
        "created_at": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        }
      }
    },
...
  }
}